New prism.Model superclass, with (very) basic model information.
Implemented by symbolic/explicit Model superclasses,
including small alignments between the two, e.g.:
- getNumStates() returns int, not long
- getNum*Initial*States(), not "Start"

Stored and made available from prism.Prism via getBuiltModel(),
which previously returned the symbolic version.

Removes a dependency on symbolic/explicit models in GUI.

+++ prism/src/strat/MDStrategyIV.java
@@ -57,7 +57,7 @@ public class MDStrategyIV extends StrategyWithStates<Double> implements MDStrate
 	public MDStrategyIV(Model model, IntegerVector iv)
 	{
 		this.model = model;
-		numStates = (int) model.getNumStates();
+		numStates = model.getNumStates();
 		actions = model.getSynchs();
 		this.iv = iv;
 		setStateLookUp(state -> {
